Fix float errors

+++ pynocchio/go_to_page_dialog.py
@@ -15,12 +15,11 @@ class GoToDialog(QtWidgets.QDialog):
         page_number = comic_handler.get_current_page().number
 
         self.model_handler = comic_handler
-        self.handler = ComicPageHandlerFactory.create_handler(
-            False, comic=comic_handler.comic, index=page_number - 1)
+        self.handler = ComicPageHandlerFactory.create_handler(False, comic=comic_handler.comic, index=page_number - 1)
 
         self.last_page = page_number
 
-        self.ui.total_page_label.setText(self.tr('of %d') % page_qty)
+        self.ui.total_page_label.setText(self.tr("of %d") % page_qty)
 
         self.ui.horizontal_slider.setMaximum(page_qty)
         self.ui.spin_box_go_page.setMaximum(page_qty)
@@ -42,8 +41,7 @@ class GoToDialog(QtWidgets.QDialog):
         self.last_page = self.ui.spin_box_go_page.value()
 
         image_page = self.handler.get_current_page_image()
-        image_page = image_page.scaledToHeight(self.height() * 0.7,
-                                               QtCore.Qt.SmoothTransformation)
+        image_page = image_page.scaledToHeight(int(self.height() * 0.7), QtCore.Qt.SmoothTransformation)
         self.ui.page_label.setPixmap(image_page)
 
         super(GoToDialog, self).update()
